Move most of the instance parsing login into lib/aws.go
