+++ cmd/test.go
@@ -21,13 +21,13 @@ Allows to identify permission issues early.
 	},
 	Run: func(cmd *cobra.Command, args []string) {
 		profiles := viper.Get("profilesConfig").([]lib.ProfileConfig)
-		summaries, err := lib.TraverseProfiles(profiles)
+		summaries, err := lib.TraverseProfiles(profiles, viper.GetBool("no-profile-prefix"))
 		if err != nil {
 			log.WithError(err).Fatal("Can't traverse through all profiles")
 		} else {
 			log.Info("All profiles have been traversted through without errors")
 			for _, summary := range summaries {
-				log.WithFields(log.Fields{"profile": summary.Name}).Infof("found %d instances", len(summary.Instances))
+				log.WithFields(log.Fields{"profile": summary.Name}).Infof("found %d instances", len(summary.SSHEntries))
 			}
 		}
 	},
